sandbox/requirements.txt
# Python packages
playwright==1.38.0
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
docker==6.1.3
aiohttp==3.8.5
websockets==11.0.3